Fix bug when drawing text with Fl_OpenGL_Graphics_Driver and scaled GUI.
The bug appeared in test/cube where the "Text" string was mispositioned
when the app was rescaled.

git-svn-id: file:///fltk/svn/fltk/branches/branch-1.4@12597 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/gl_draw.cxx
@@ -377,7 +377,7 @@ void gl_draw_image(const uchar* b, int x, int y, int w, int h, int d, int ld) {
  but not text font, and only ASCII characters can be drawn.
  */
 
-static int gl_scale = 1; // set to 2 for high resolution Fl_Gl_Window
+static float gl_scale = 1; // set to 2 for high resolution Fl_Gl_Window
 static int has_texture_rectangle = 0; // true means GL_EXT_texture_rectangle is available
 
 #include <FL/glu.h>  // for gluUnProject() and gluCheckExtension()
